+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Improves Windows build support by including binary distribution of GEOS
 * Embeds GEOS 3.5.0-1 on Windows
+* Change "Portability" to "portable"
 
 # v0.1.5.0